+++ tests/__main__.py
@@ -1,1 +1,1 @@
-from tests import convert, rules, models
+from tests import convert, rules, models, io